requirements.txt
flask
azure-ai-formrecognizer==3.3.0
python-dotenv
Werkzeug
